restructure asab-config folder
